packages/artifact_schema/pyproject.toml
[project]
name = "artifact-schema"
version = "0.1.0"
description = "Pydantic contract for BenchmarkAgent artifacts"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/artifact_schema"]
